Added handling of fpu exceptions. It's a build configuration option, so it can be disabled. It only catches fpu exceptions in the machines' work functions, but this could be extended to other functions as well. Each exception type is reported only once per machine it occurs, by growing the fpu exception mask. "Inexact result" fpu exceptions are initially already masked.
Also, made exceptions which happen in machines appear not only in the lof, but also in a message box, because many people wondered why some machines were automagically set to bypassed/muted (The message box wouldn't be needed if the MFC logging window worked.)

+++ trunk/portable_psycle/src/psycle/plugins/crasher.cpp
@@ -36,8 +36,11 @@ protected:
 		//throw "crash on purpose!";
 		//throw std::runtime_error("crash on purpose!");
 
-		// division by 0:
-		volatile int i(0); i = 0 / i; // trick so that the compiler does not remove the code when optimizing
+		// integer division by 0:
+		//volatile int i(0); i = 0 / i; // trick so that the compiler does not remove the code when optimizing
+
+		// floating point division by 0:
+		volatile float i(0); i = 0 / i; // trick so that the compiler does not remove the code when optimizing
 
 		// infinite loop so that we can test interruption signal:
 		//while(true);
